+++ app/Exceptions/Handler.php
@@ -43,8 +43,7 @@ class Handler extends ExceptionHandler
      */
     public function register()
     {
-        $this->reportable(function () {
-            //
+        $this->reportable(function (Throwable $e) {
         });
     }
 }